+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog of lizard-wbconfiguration
 ===================================================
 
 
-0.1.4 (unreleased)
+0.1.4 (2012-01-23)
 ------------------
 
 - Fixed problem of repeated columns in dbf (issue #3).